Added method logic(int[]) for solve Example1 Refactoring: 1) renamed class Main -> ArrayExample1 2) method fillArray() is placed in separate class Utils
